[FIX] Fix MetaResult (#154)
* Move MetaResult from base.meta to base.base.

* Fix tests.

* Remove empty tests.

* Fix MetaResult calls.

* Fix IBMAEstimator.
