+++ nimare/base/__init__.py
@@ -2,13 +2,17 @@
 Top-level namespace for nimare base.
 """
 
+from .base import MetaResult
 from .decode import Decoder
-from .meta import (KernelTransformer, MetaEstimator, CBMAEstimator,
-                   IBMAEstimator, MetaResult)
-from .annotate import (AnnotationModel)
-from .misc import (Parcellator)
+from .meta import KernelTransformer, CBMAEstimator, IBMAEstimator
+from .annotate import AnnotationModel
+from .misc import Parcellator
 
 __all__ = ['Decoder',
-           'KernelTransformer', 'MetaEstimator', 'CBMAEstimator',
-           'IBMAEstimator', 'MetaResult',
-           'AnnotationModel', 'Parcellator']
+           'KernelTransformer',
+           'CBMAEstimator',
+           'IBMAEstimator',
+           'MetaResult',
+           'AnnotationModel',
+           'Parcellator'
+           ]
