--- nimare/meta/cbma/tests/test_kernel.py
@@ -1,15 +1,0 @@
-"""
-Tests for nimare.meta.cbma.kernel
-"""
-
-
-def test_alekernel():
-    pass
-
-
-def test_mkdakernel():
-    pass
-
-
-def test_kdakernel():
-    pass
