+++ nimare/tests/test_cbma_ale.py
@@ -46,7 +46,7 @@ def test_ale(testdata1):
     """
     ale_meta = ale.ALE(testdata1)
     ale_meta.fit(n_iters=5, ids=testdata1.ids, n_cores=1)
-    assert isinstance(ale_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(ale_meta.results, nimare.base.base.MetaResult)
 
 
 def test_ale_subtraction(testdata1):
@@ -56,4 +56,4 @@ def test_ale_subtraction(testdata1):
     ale_meta = ale.ALE(testdata1)
     ale_meta.fit(n_iters=5, ids=testdata1.ids[:5], ids2=testdata1.ids[5:],
                  n_cores=1)
-    assert isinstance(ale_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(ale_meta.results, nimare.base.base.MetaResult)
