+++ nimare/tests/test_cbma_mkda.py
@@ -46,7 +46,7 @@ def test_mkda_density(testdata1):
     """
     mkda_meta = mkda.MKDADensity(testdata1)
     mkda_meta.fit(n_iters=5, ids=testdata1.ids, n_cores=1)
-    assert isinstance(mkda_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(mkda_meta.results, nimare.base.base.MetaResult)
 
 
 def test_mkda_chi2_fdr(testdata1):
@@ -55,7 +55,7 @@ def test_mkda_chi2_fdr(testdata1):
     """
     mkda_meta = mkda.MKDAChi2(testdata1)
     mkda_meta.fit(corr='fdr', ids=testdata1.ids, ids2=testdata1.ids, n_cores=1)
-    assert isinstance(mkda_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(mkda_meta.results, nimare.base.base.MetaResult)
 
 
 def test_mkda_chi2_fwe(testdata1):
@@ -65,7 +65,7 @@ def test_mkda_chi2_fwe(testdata1):
     mkda_meta = mkda.MKDAChi2(testdata1)
     mkda_meta.fit(n_iters=5, ids=testdata1.ids, ids2=testdata1.ids, corr='fwe',
                   n_cores=1)
-    assert isinstance(mkda_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(mkda_meta.results, nimare.base.base.MetaResult)
 
 
 def test_kda_density(testdata1):
@@ -74,4 +74,4 @@ def test_kda_density(testdata1):
     """
     kda_meta = mkda.KDA(testdata1)
     kda_meta.fit(n_iters=5, ids=testdata1.ids, n_cores=1)
-    assert isinstance(kda_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(kda_meta.results, nimare.base.base.MetaResult)
