+++ nimare/tests/test_cbma_scale.py
@@ -51,4 +51,4 @@ def test_scale(testdata1):
     ijk = ijk[:, :100]
     scale_meta = ale.SCALE(testdata1, ijk=ijk)
     scale_meta.fit(n_iters=5, ids=testdata1.ids, n_cores=1)
-    assert isinstance(scale_meta.results, nimare.base.meta.MetaResult)
+    assert isinstance(scale_meta.results, nimare.base.base.MetaResult)
